% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{benefitIntegral}
\alias{benefitIntegral}
\title{Benefit integral}
\usage{
benefitIntegral(frac, par)
}
\arguments{
\item{frac}{Fraction of population that has been vaccinated by the end of the period}

\item{par}{Parameters object with model parameters}
}
\value{
Share of benefits obtained
}
\description{
Computes the share of benefits obtained when a fraction `frac` of the population is vaccinated by the end of the analysis period.
}
