% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{benefitIntegralDisc}
\alias{benefitIntegralDisc}
\title{Benefit integral with discounting}
\usage{
benefitIntegralDisc(frac1, frac2, par, share1 = 1, share2 = 1)
}
\arguments{
\item{frac1}{Lower limit of integral}

\item{frac2}{Upper limit of integral}

\item{par}{Parameters object}

\item{share1}{Share of damage at time of frac1}

\item{share2}{Share of damage at time of frac2}
}
\value{
Values of benefits integral
}
\description{
More general function that allows for linear damage discounting due to introduction of treatment
}
\examples{
par = Parameters$new(benefitdist='pnorm',alpha=1)
benefit = benefitIntegralDisc(0,0.5,par)
}
