% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{benefits}
\alias{benefits}
\title{Benefits from vaccination}
\usage{
benefits(monthben, capacities, endtimes, par)
}
\arguments{
\item{monthben}{Benefits, i.e., total harm (economic + health) due to covid}

\item{capacities}{List of vectors with capacities. Each vector represents a period. Elements
in the vector represent the capacities for that period in each scenario.}

\item{endtimes}{Vector of endtimes of all the periods}

\item{par}{\code{Parameters} object with model parameters}
}
\value{
Total benefits from vaccination
}
\description{
Computes the benefits obtained from vaccination by integrating the share of harm reduction over time.
}
\examples{
cap = list(c(0.1,0.2,0.4),c(0.2,0.5,0.7))
par = Parameters$new()
ben = benefits(100, cap, c(3,4),par)
}
