% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{benefitsOld}
\alias{benefitsOld}
\title{Benefits from vaccination}
\usage{
benefitsOld(
  monthben,
  capacity,
  endtime,
  par,
  capacity2 = NULL,
  endtime2 = NULL
)
}
\arguments{
\item{monthben}{Benefits, i.e., total harm (economic + health) due to covid}

\item{capacity}{Vaccination capacity (billions per month)}

\item{endtime}{Time at which vaccination ends}

\item{par}{`Parameters` object with model parameters}

\item{capacity2}{Second capacity after endtime (billions per month)}

\item{endtime2}{Time at which vaccination with capacity2 ends}
}
\value{
Total benefits from vaccination
}
\description{
Computes the benefits obtained from vaccination by integrating the share of harm reduction over time.
}
