% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{candidateDraws}
\alias{candidateDraws}
\title{Candidate success draws}
\usage{
candidateDraws(dcandidate, par, seed = 30)
}
\arguments{
\item{dcandidate}{\code{data.table} with information about candidates}

\item{par}{\code{Parameters} object with model parameters}

\item{seed}{Random seed}
}
\value{
\code{data.table} with a summary of the candidates
}
\description{
Generate Monte Carlo draws for the success and failure of candidates
}
