% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{candidatesFung}
\alias{candidatesFung}
\title{Candidates with fungible costs}
\usage{
candidatesFung(d, par, computeExpComp = F, seed = 10)
}
\arguments{
\item{d}{Summary data with candidates}

\item{par}{Parameters object with model parameters}

\item{computeExpComp}{Whether to compute expected competition for all candidates for all numbers of potential competitors}

\item{seed}{Seed for random generator}
}
\value{
List with two data.tables. \code{dordered} lists all the candidates in the optimal order and with some statistics.
\code{dcanddraws} includes all the random draws and their outcomes.
}
\description{
Based on summary data for candidates, finds the optimal order to select those candidades, as well as some statistics about
them. Some of those statistics are computed by a Monte Carlo simulation.
}
