% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{capacityCandidatesCombinations}
\alias{capacityCandidatesCombinations}
\title{Combinations of capacity and candidates}
\usage{
capacityCandidatesCombinations(dordered, par, prodGrid = 0.5, maxprodfac = 4)
}
\arguments{
\item{dordered}{data.table with ordered candidates}

\item{par}{Parameters object with model parameters}

\item{prodGrid}{Resolution of the grid of production capacities}

\item{maxprodfac}{Maximum capacity, as a factor of the capacity constraint}
}
\value{
data.table with all combinations
}
\description{
Computes social benefits, costs, and program cost for different combinations of the
number of candidates and capacity per candidate
}
