% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{computeCandidateTable}
\alias{computeCandidateTable}
\title{Compute optimal candidate table}
\usage{
computeCandidateTable(d, par, computeExpComp = F)
}
\arguments{
\item{d}{data.table with summarized data}

\item{par}{Parameters object with all parameters for the model}

\item{computeExpComp}{}
}
\value{
data.table with the optimal candidates ordered by the order in which they should be chosen
}
\description{
Computes the optimal ordering of candidates to maximize the probability of getting at least one vaccine.
Also computes a few additional statistics that are necessary for rest of the model.
}
