% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{countryExpectedBenefits}
\alias{countryExpectedBenefits}
\title{Expected benefits for one country}
\usage{
countryExpectedBenefits(
  capacities,
  dcandidate,
  targetPermutations,
  dplatforms,
  par,
  grid = 1
)
}
\arguments{
\item{capacities}{Vector of capacities for each candidate}

\item{dcandidate}{`data.table` with candidate information}

\item{targetPermutations}{`data.table` with permutations of targets}

\item{dplatforms}{`data.table` with platform information}

\item{par}{`Parameters` object with model parameters}

\item{grid}{Size of the capacity grid}
}
\value{
Expected benefits from vaccination
}
\description{
Compute expected benefits from a portfolio for one country
}
