% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{countryNetBenefits}
\alias{countryNetBenefits}
\title{Net benefits for a country}
\usage{
countryNetBenefits(
  capacities,
  dcandidate,
  targetPermutations,
  dplatforms,
  step,
  price,
  par,
  lambda = 1
)
}
\arguments{
\item{capacities}{Vector of capacities for each candidate}

\item{dcandidate}{`data.table` with candidate information}

\item{targetPermutations}{`data.table` with permutations of targets}

\item{dplatforms}{`data.table` with platform information}

\item{step}{Size of the capacity grid}

\item{price}{Price per vaccine / year}

\item{par}{`Parameters` object with model parameters}

\item{lambda}{Lagrange multiplier. Should be one, unless trying to solve problem with constrained budget.}
}
\value{
Expected benefits for the country
}
\description{
Expected net benefits for a country that invests in a certain portfolio
}
