% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{countryParameters}
\alias{countryParameters}
\title{Country parameters}
\usage{
countryParameters(countryData)
}
\arguments{
\item{countryData}{\code{data.table} with country data}
}
\value{
List of \code{Parameters} objects for all countries in the dataset
}
\description{
Get a list of \code{Parameters} objects for all countries from country data
}
\examples{
countryData = loadCountryData(system.file('extdata',
'countryData.xlsx',package = 'vaccineEarlyInvest'))
countryPar = countryParameters(countryData)
}
