% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{crossJoin}
\alias{crossJoin}
\title{Cross join}
\usage{
crossJoin(db1, db2)
}
\arguments{
\item{db1}{First dataset}

\item{db2}{Second dataset}
}
\value{
Cartesian product of both datasets
}
\description{
Cartesian product of two data.tables
}
