% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{demandPriceTaker}
\alias{demandPriceTaker}
\title{Title}
\usage{
demandPriceTaker(
  population,
  gdp_pc,
  frac_high_risk,
  loss2yr,
  price,
  inisteps = c(10, 1, 0.1),
  mainstep = 0.1,
  prices = seq(100, 1, -1),
  candidateFile = NULL
)
}
\arguments{
\item{candidateFile}{}
}
\value{

}
\description{
Title
}
