% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{expectedBenefitsTable}
\alias{expectedBenefitsTable}
\title{Expected benefits}
\usage{
expectedBenefitsTable(distribution, benefitsTable, grid = 1)
}
\arguments{
\item{distribution}{\code{data.table} with the distribution of effective capacity}

\item{benefitsTable}{\code{data.table} with benefits as a function of effective capacity}

\item{grid}{Resolution of the capacity grid}
}
\value{
Expected benefits
}
\description{
Computes expected benefits from some capacity distribution and from a benefits table
}
