% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fundingStructure}
\alias{fundingStructure}
\title{Funding structure}
\usage{
fundingStructure(data, opt, par)
}
\arguments{
\item{data}{List with output from candidatesFung()}

\item{opt}{List with output from optimizaitonsFung()}

\item{par}{Parameters object with model parameters}
}
\value{

}
\description{
Computes characteristics of the funding structure, including the minimum pull amount that satisfies candidates'
participation constraint, every candidate's expected earnings, and the price they would have to expect to prefer to
choose an outside market instead of the program.
}
