% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{getTargetPermutations}
\alias{getTargetPermutations}
\title{Target permutations}
\usage{
getTargetPermutations(targets, probs)
}
\arguments{
\item{targets}{vector with target names}

\item{probs}{vector with target probabilities}
}
\value{
data.table with all permutations and their probabilities
}
\description{
Create data.table with information about all possible permutations of target success/failure
}
