% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{globalNetBenefits}
\alias{globalNetBenefits}
\title{Title}
\usage{
globalNetBenefits(
  capacities,
  dcandidate,
  targetPermutations,
  dplatforms,
  benefitsTable,
  par,
  grid = 1
)
}
\arguments{
\item{dcandidate}{}
}
\value{

}
\description{
Title
}
