% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{limBudgetPriceTaker}
\alias{limBudgetPriceTaker}
\title{Optimal portfolio with limited budget}
\usage{
limBudgetPriceTaker(
  parameters = NULL,
  population,
  gdp_pc,
  frac_high_risk,
  loss2yr,
  price,
  shadowCosts = seq(1, 5, 0.1),
  inisteps = c(10, 1),
  mainstep = 0.1,
  candidateFile = NULL,
  verbose = 0
)
}
\arguments{
\item{parameters}{\code{Parameters} object with model parameters}

\item{population}{Country population (in millions)}

\item{gdp_pc}{Country GDP per capita (in thousand $)}

\item{frac_high_risk}{Fraction of population that is high risk}

\item{loss2yr}{Cumulative percent of GDP lost because of pandemic over two years}

\item{price}{Market price for manufacturing capacity ($ per course / year)}

\item{shadowCosts}{Vector with shadow cost of public funds}

\item{inisteps}{Step sizes to optimize over for the initial point}

\item{mainstep}{Step size for main optimization}

\item{candidateFile}{File with candidate data}

\item{verbose}{How much output to produce. 0 means no output, 1 means limited output, 2 means full output.}
}
\value{
List with information on demand curve. Includes a \code{data.table} with total demand, benefits, and cost,
and a matrix with demand for individual candidates at every price
}
\description{
Computes optimal portfolios for a price-taking country when there is a limited budget at
different levels
}
\examples{
\dontrun{
population <- 31.99
gdp_pc <- 6.71
frac_high_risk <- 0.131
loss2yr <- 0.269
par <- Parameters$new(population=population, gdp_pc=gdp_pc,
                   frac_high_risk=frac_high_risk,
                   loss2yr=loss2yr)
                   demand <- demandPriceTaker(parameters=par)
}
}
