% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{loadCountryData}
\alias{loadCountryData}
\title{Load data from countries}
\usage{
loadCountryData(filename, Gavi = F)
}
\arguments{
\item{filename}{File name (with path) of the xlsx file with country data}

\item{Gavi}{Logical, whether to treat Gavi countries as a blcok}
}
\value{
a \code{data.table} with country data
}
\description{
Loads a .xlsx file with country data, including demographics, GDP, and economic impact due to Covid-19
}
\examples{
countrydata <- loadCountryData('inst/extdata/countryData.xlsx')
}
