% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{loadData}
\alias{loadData}
\title{Load data}
\usage{
loadData(par, candidateFile = NULL, includeVaccines = c())
}
\arguments{
\item{par}{\code{Parameters} object with model parameters}

\item{candidateFile}{String with file name (including path) for candidate information}

\item{includeVaccines}{List of experimental vaccines to include in analysis}
}
\value{
Data.table with summary data
}
\description{
Load summary data with characteristics for vaccines
}
\examples{
par=Parameters$new()
d = loadData(par, includeVaccines = 'BCG')
d = loadData(par,candidatefile = 
             "inst/extdata/vaccinesSummary.csv")
}
