% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{optBudget}
\alias{optBudget}
\title{Optimal policy under a budget constraint}
\usage{
optBudget(dordered, dcanddraws, par)
}
\arguments{
\item{dordered}{Ordered dataset with candidate vaccines}

\item{dcanddraws}{Monte Carlo draws of vaccines}

\item{par}{Parameter object with model parametesr}
}
\value{
A list with all the potential combinations of candidates and capacity, and a list with the optimal policiy for
different levels of the budget
}
\description{
Compute the optimal policy when there is a budget constraint. It computes the optima both with and without
a capacity constraint.
}
