% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{optCandidates}
\alias{optCandidates}
\title{Optimal number of candidates}
\usage{
optCandidates(dordered, par)
}
\arguments{
\item{par}{Parameters object with all parameters for the model}

\item{data}{data.table with ordered candidates}
}
\value{
List with the optimal number of candidates, the marginal probability at the optimum, the marginal
cost at the optimum, and the benefit of finding a vaccine
}
\description{
Based on a dataset of candidates ordered optimally, compute optimal number of candidates as well as a few other numbers.
}
