% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{optCandidatesCapacity}
\alias{optCandidatesCapacity}
\title{Optimal number of candidates}
\usage{
optCandidatesCapacity(data, par)
}
\arguments{
\item{data}{data.table with ordered candidates}

\item{par}{Parameters object with model parameters}
}
\value{
List with the optimal number of candidates, the marginal probability at the optimum, the marginal
cost at the optimum, and the benefit of finding a vaccine
}
\description{
Based on a dataset of candidates ordered optimally, compute optimal number of candidates as well as a few other numbers.
}
