% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{optimizationsFung}
\alias{optimizationsFung}
\title{Optimiazation with fungible costs}
\usage{
optimizationsFung(dordered, dcanddraws, par)
}
\arguments{
\item{dordered}{`data.table` with candidates ordered in the optimal selection order.}

\item{dcanddraws}{`data.table` with Monte Carlo draws}

\item{par}{Parameters object with model parameters}
}
\value{
List with several objects related to all the optimization that were performed
}
\description{
Computes the optimal policy assuming fungible costs. Also computes the optimal policy with a limited budget.
}
