% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{optimizeGrid}
\alias{optimizeGrid}
\title{Maximize objective function staying within a grid}
\usage{
optimizeGrid(capacities, objectiveFun, step = 100, verbose = 1, ...)
}
\arguments{
\item{capacities}{Starting point}

\item{objectiveFun}{Function to optimize}

\item{step}{Step size of the grid}

\item{verbose}{Whether to print extensive output (2), some output (1), or no output (0)}

\item{...}{Other parameters to pass on to the objective function}
}
\value{
Optimal capacities
}
\description{
Implements an algorithm to find the optimum of the objective function without ever moving outside of a grid with
a fixed step size.
}
