% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{overallDistribution}
\alias{overallDistribution}
\title{Overall distribution of capacity}
\usage{
overallDistribution(
  dcandidate,
  targetPermutations,
  dplatforms,
  poverall,
  psubcat,
  grid = 1,
  target = T
)
}
\arguments{
\item{dcandidate}{data.table with information by candidate}

\item{targetPermutations}{data.table with information about target permutations}

\item{dplatforms}{data.table with information by platform}

\item{poverall}{Overall probability that a vaccine is feasible}

\item{psubcat}{Probability that no problem shows up at the subcategory level}

\item{grid}{Resolution of the grid to compute all distributions}

\item{target}{Whether to allow for target distributions (otherwise all targets succeed with probability 1)}
}
\value{
data.table with the overall distribution
}
\description{
Computes the distribution of total capacity based on data by candidate and platform
}
