% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\name{Parameters}
\alias{Parameters}
\title{Model parameters}
\description{
Class to handle parameters of the model for investment in early vaccine capacity
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{poverall}}{Probability that no problem at the overall level prevents vaccine feasibility}

\item{\code{pvector}}{Probability that there's no problem at the viral vector platform level}

\item{\code{psubunit}}{Probability that there's no problem at the protein subunit platform level}

\item{\code{prna}}{Probability that there's no problem at the RNA platform level}

\item{\code{pdna}}{Probability that there's no problem at the DNA platform level}

\item{\code{pattenuated}}{Probability that there's no problem at the live attenuated platform level}

\item{\code{pinactivated}}{Probability that there's no problem at the inactivated platform level}

\item{\code{pvlp}}{Probability that there's no problem at the virus-like particle platform level}

\item{\code{ppreclinical}}{Probability that there's no problem at the candidate level when a vaccine is in preclincal trials}

\item{\code{plivebac}}{Probability that there's no problem at the live attenuated bacteria platform level}

\item{\code{paapc}}{Probability that there's no problem at the artificial antigen presenting cells platform level}

\item{\code{pdendritic}}{Probability that there's no problem at the dendritic cells platform level}

\item{\code{psav}}{Probability that there's no problem at the self-assembling vaccine platform level}

\item{\code{punknown}}{Probability that there's no problem for unknown platform}

\item{\code{prepurposed}}{Probability that there's no problem with a repurposed candidate}

\item{\code{pphase1}}{Probability that there's no problem at the candidate level when a vaccine is in phase 1 trials}

\item{\code{pphase2}}{Probability that there's no problem at the candidate level when a vaccine is in phase 2 trials}

\item{\code{pphase3}}{Probability that there's no problem at the candidate level when a vaccine is in phase 2 trials}

\item{\code{psubcat}}{Probability that there's no problem at subcategory level}

\item{\code{pspike}}{Probability that there's no problem at the candidate level when a vaccine targets spike proten}

\item{\code{precombinant}}{Probability that there's no problem at the candidate level when a vaccine targets recombinant proten}

\item{\code{potherprotein}}{Probability that there's no problem at the candidate level when a vaccine targets some other proten}

\item{\code{popshare}}{Share of world population accounted for by country / coalition}

\item{\code{gdpshare}}{Share of world GDP accounted for by country /coalition}

\item{\code{econlossratio}}{Ratio of economic losses in the country / coalition as a percent relative to world}

\item{\code{spillovers}}{How much do countries in the country /coalition care about spillovers}

\item{\code{outlifefactor}}{Extent to which countries care about international lives}

\item{\code{outlivesratio}}{Fraction of world that lives outside country /coalition}

\item{\code{fracHighRisk}}{Fraction of population in the country /coalition at high risk}

\item{\code{TT}}{Time that the vaccine is brought forward (months)}

\item{\code{tau}}{Period of analysis (months)}

\item{\code{damageshare}}{Fraction of damage accounted for by some fraction of the population}

\item{\code{vaccshare}}{Fraction of population that accounts for some fraction of damage}

\item{\code{monthben}}{Economic benefit for the whole world per month (in billion $)}

\item{\code{worldmortality}}{World mortality}

\item{\code{mortality}}{Monthly internal mortality}

\item{\code{statLife}}{Value of statistical life in million $}

\item{\code{lifeExp}}{Life expectancy}

\item{\code{yearsLost}}{Years lost per covid death}

\item{\code{sharm}}{Expected share of harm avoided due to an alternative drug being developed}

\item{\code{maxcand}}{Number of candidates to consider}

\item{\code{hben}}{Monthly health benefits}

\item{\code{damageint}}{Integral that measures fraction of damages}

\item{\code{totmonthben}}{Total benefit of bringing forward vaccine for one month}

\item{\code{alpha}}{Parameter for heterogeneity of benefits of vaccination with a smooth functional form.
a=1 corresponds to perfectly homogeneous benefits, a=0 corresponds to all benefits to the first person.}

\item{\code{benefitdist}}{String describing the form of the benefit distribution}

\item{\code{piecewisepar}}{Parameters of a piecewise linear benefit distribution}

\item{\code{fracneeded}}{Fraction of population needed to reopen economy}

\item{\code{effpop}}{Effective world population that needs to be vaccinated to
get all (100 percent) of benefits}

\item{\code{c}}{Cost of capacity per dose / year}

\item{\code{capacity}}{Constraint on capacity (doses / month)}

\item{\code{capkink}}{Capacity at which marginal cost has a kink (doses / month)}

\item{\code{capkink_nucleic}}{Capacity at which marginal cost has a kink for RNA and DNA (doses / month)}

\item{\code{capkink_subunits}}{Capacity at which marginal cost has a kink for protein subunit and viral vector (doses / month)}

\item{\code{capkink_live}}{Capacity at which marginal cost has a kink for live attenuated and deactivated (doses / month)}

\item{\code{mgcostelast}}{Elasticity of marginal cost after kink}

\item{\code{afterCapacity}}{Worldwide capacity after the end of the program (doses / month)}

\item{\code{counterCapacity}}{Worldwide capacity in the counterfactual (doses / month)}

\item{\code{pop}}{World population population (billions)}

\item{\code{cprod}}{Production cost per vaccine ($)}

\item{\code{replications}}{Number of replications for Monte Carlo}

\item{\code{fcapacity}}{Fraction of capacity cost that must be borne by the firm}

\item{\code{fcostred}}{Fraction of cost reduction after the second vaccine within a platform}

\item{\code{outsidefactor}}{How much more likely are firms to be successful in an outside market}

\item{\code{pearly}}{Price of early batch of vaccines}

\item{\code{plate}}{Price of late batch of vaccines}

\item{\code{nearly}}{Number or vaccines in early batch}

\item{\code{nlate}}{Number or vaccines in late batch}

\item{\code{riskPremium}}{Premium that must given to candidates to participate despite risk (bn $)}

\item{\code{overallFung}}{Fraction of capacity cost that is fungible across platforms}

\item{\code{platFung}}{Fraction of capacity cost that is fungible within platforms}

\item{\code{subcatFung}}{Fraction of capacity cost that is fungible within subcategory}

\item{\code{protVectorFung}}{Fraction of capacity cost that is fungible between protein and viral vectors}

\item{\code{fracScrap}}{Fraction of cost that can be saved if a candidate is not successful}

\item{\code{budget}}{Program budget}

\item{\code{capcost}}{Total capacity cost per candidate}

\item{\code{global}}{Whether the the analysis is for the whole wolrd}

\item{\code{outcap}}{Capacity to build in an outside market}

\item{\code{outprob}}{Probability that copy vaccine will be successful, conditional on own success}

\item{\code{outquantity}}{Amount of outside vaccines that will be sold at high prices}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Parameters$new()}}
\item \href{#method-setDerivedParameters}{\code{Parameters$setDerivedParameters()}}
\item \href{#method-clone}{\code{Parameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Parameters} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameters$new(
  input = NULL,
  population = NULL,
  gdp_pc = NULL,
  frac_high_risk = NULL,
  loss2yr = NULL,
  worldLoss2yr = 0.138,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{Determines how to initialize object. Can be a string telling which default parameters to use. Can also be
the \code{input} object (of class \code{reactivevalues}) with the inputs from a shiny app, in which case all inputs are copied into
fields.}

\item{\code{population}}{Country population (in millions)}

\item{\code{gdp_pc}}{Country GDP per capita (in thousand $)}

\item{\code{frac_high_risk}}{Fraction of population that is high risk}

\item{\code{loss2yr}}{Cumulative percent of GDP lost because of pandemic over two years}

\item{\code{worldLoss2yr}}{Cumulative percent of GDP lost worldwide because of pandemic over two years}

\item{\code{...}}{Set fields at non-default values}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Parameters} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setDerivedParameters"></a>}}
\if{latex}{\out{\hypertarget{method-setDerivedParameters}{}}}
\subsection{Method \code{setDerivedParameters()}}{
Compute additional parameters that are functions of the input parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameters$setDerivedParameters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
