% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{platformDistribution}
\alias{platformDistribution}
\title{Distribution of capacity in a platform}
\usage{
platformDistribution(plat, subcatDists, psubcat)
}
\arguments{
\item{plat}{Character with the subcategory name}

\item{subcatDists}{Data.table with the distributions for every subcategory in the platform}
}
\value{
A data.table summarizing the distribution of total capacity in the platfom
}
\description{
Finds the distribution of the total capacity in a platform depending on distributions for subcategories in
the platform
}
