% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{portfolioPriceTaker}
\alias{portfolioPriceTaker}
\title{Optimal portfolio for a price taker}
\usage{
portfolioPriceTaker(
  population,
  gdp_pc,
  frac_high_risk,
  loss2yr,
  price,
  steps = c(10, 1, 0.1),
  candidateFile = NULL
)
}
\arguments{
\item{population}{Country population (in millions)}

\item{gdp_pc}{Country GDP per capita (in thousand $)}

\item{frac_high_risk}{Fraction of population that is high risk}

\item{loss2yr}{Cumulative percent of GDP lost because of pandemic over two years}

\item{price}{Market price for capacity ($ per course / year)}

\item{steps}{Steps to optimize over}

\item{candidateFile}{File with candidate data}
}
\value{
Vector with optimal capacities
}
\description{
Computes the optimal portfolio for a price-taking country
}
