% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{priceTakerCost}
\alias{priceTakerCost}
\title{Cost for a price taker}
\usage{
priceTakerCost(capacities, price)
}
\arguments{
\item{capacities}{Vector of capacities for all candidates}

\item{price}{Price per vaccine / year}
}
\value{
Total cost of the portfolio
}
\description{
Computes the total cost of a portfolio for a player that is a price taker
}
