% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{progStructure}
\alias{progStructure}
\title{Program structure}
\usage{
progStructure(dordered, dcanddraws, optCand, par)
}
\arguments{
\item{dordered}{data.table with candidates with the optimal order}

\item{dcanddraws}{data.table with draws from Monte Carlo simulation}

\item{optCand}{Optimal number of candidates}

\item{par}{Parameters object with model parameters}
}
\value{
List with pull and push amounts, cost of capacity, intercept of the pull quantity, slope of the pull quantity,
optimal number of candidates, and maximum push amount
}
\description{
Computes some characteristics of the optimal program
}
