% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{socialCost}
\alias{socialCost}
\title{Social cost}
\usage{
socialCost(capacities, distribution, par)
}
\arguments{
\item{capacities}{Vector with capacities for all candidates}

\item{distribution}{`data.table` with the distribution of capacities}

\item{par}{`Parameters` object with main model parameters}
}
\value{
Social cost of the portfolio
}
\description{
Social cost of a portfolio
}
