% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{subcatDistribution}
\alias{subcatDistribution}
\title{Distribution of capacity in a subcategory}
\usage{
subcatDistribution(sub, dcandidate)
}
\arguments{
\item{sub}{Character with the subcategory name}

\item{dcandidate}{Data.table with information by candidate}
}
\value{
A data.table summarizing the distribution of total capacity in the subcategory
}
\description{
Finds the distribution of the total capacity in a subcategory depending on individual candidate characteristics
}
