% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{sumdist}
\alias{sumdist}
\title{Sum of distributions}
\usage{
sumdist(dist1, dist2)
}
\arguments{
\item{dist1}{First distribution. Vector where position i corresponds to the probability that the probability is i-1}

\item{dist2}{Second distribution. Vector where position i corresponds to the probability that the probability is i-1}
}
\value{
DIstribution of the sum. Vector where position i corresponds to the probability that the probability is i-1
}
\description{
Distribution of the sum of two indepenent discrete random variables
}
