% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{sumdistSelf}
\alias{sumdistSelf}
\title{Sum of distribution with itself}
\usage{
sumdistSelf(dist, N = 2)
}
\arguments{
\item{dist}{Input distribution. Vector where position i corresponds to the probability that the probability is i-1}

\item{N}{Number of times to add}
}
\value{
Distribution of the sum. Vector where position i corresponds to the probability that the probability is i-1
}
\description{
Computes the sum of the distribution of N iid discrete random variables
}
