% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{countryNetBenefits}
\alias{countryNetBenefits}
\title{Net benefits for a country}
\usage{
countryNetBenefits(
  capacities,
  dcandidate,
  targetPermutations,
  dplatforms,
  grid,
  price,
  par,
  lambda = 1
)
}
\arguments{
\item{capacities}{Vector of capacities for each candidate}

\item{dcandidate}{\code{data.table} with candidate information, typically from
\code{\link[=candidatesFung]{candidatesFung()}}}

\item{targetPermutations}{\code{data.table} with permutations of targets}

\item{dplatforms}{\code{data.table} with platform information}

\item{grid}{Size of the capacity grid}

\item{price}{Price per vaccine / year}

\item{par}{\code{Parameters} object with model parameters}

\item{lambda}{Lagrange multiplier. Should be one, unless trying to solve
problem with constrained budget.}
}
\value{
Expected benefits for the country
}
\description{
Expected net benefits for a country that invests in a certain portfolio.
This function is typically called by \code{\link[=portfolioPriceTaker]{portfolioPriceTaker()}}, which contains
pre-processing of relevant inputs.
}
