% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{getBenefitsTable}
\alias{getBenefitsTable}
\title{Get benefits table}
\usage{
getBenefitsTable(countryData, max = 1000, grid = 1)
}
\arguments{
\item{countryData}{\code{data.table} with country data}

\item{max}{Maximum effective capacity to compute benefits for}

\item{grid}{Resolution of the capacity grid}
}
\value{
\code{data.table} keyed by \code{capacity}. The \code{socialBenefits} column gives the benefit from the program
}
\description{
Create a table that tells the global benefits corresponding to every level of effective capacity
}
\examples{
countryData = loadCountryData(system.file('extdata',
'countryData.xlsx',package = 'vaccineEarlyInvest'))
benefitsTable = getBenefitsTable(countryData)
}
