% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{permutationDistribution}
\alias{permutationDistribution}
\title{Total capacity distribution for one permutation of platforms}
\usage{
permutationDistribution(dcandidate, dplatforms, poverall, psubcat)
}
\arguments{
\item{dcandidate}{data.table with information by candidate}

\item{dplatforms}{data.table with information by platform}
}
\value{
data.table with the distribution of total capacity
}
\description{
Compute the distribution of total capacity across candidates assuming one particular outcome for target successes
}
