% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{portfolioPriceTaker}
\alias{portfolioPriceTaker}
\title{Optimal portfolio for a price taker}
\usage{
portfolioPriceTaker(
  parameters = NULL,
  population,
  gdp_pc,
  frac_high_risk,
  loss2yr,
  price,
  steps = c(10, 1, 0.1),
  candidateFile = NULL,
  lambda = 1
)
}
\arguments{
\item{parameters}{\code{Parameters} object with model parameters}

\item{population}{Country population (in millions)}

\item{gdp_pc}{Country GDP per capita (in thousand $)}

\item{frac_high_risk}{Fraction of population that is high risk}

\item{loss2yr}{Cumulative percent of GDP lost because of pandemic over two years}

\item{price}{Market price for manufacturing capacity ($ per course / year)}

\item{steps}{Steps to optimize over}

\item{candidateFile}{File with candidate data}
}
\value{
List with information on optimal portfolio. Includes investment
in every candidate, benefits, cost, and total capacity.
}
\description{
Computes the optimal portfolio for a price-taking country.
This is done by calling optimisation function \code{\link[=optimizeGrid]{optimizeGrid()}}
on benefits function \code{\link[=countryNetBenefits]{countryNetBenefits()}}
}
\examples{
population <- 31.99 
gdp_pc <- 6.71
frac_high_risk <- 0.131
loss2yr <- 0.269
price <- 4
portfolio <- portfolioPriceTaker(population=population, gdp_pc=gdp_pc,
                   frac_high_risk=frac_high_risk, 
                   loss2yr=loss2yr, price=price,
                   candidateFile="../inst/extdata/vaccinesSummary.csv")
or 

par <- Parameters$new(population=population, gdp_pc=gdp_pc,
                   frac_high_risk=frac_high_risk, 
                   loss2yr=loss2yr)
portfolio <- portfolioPriceTaker(parameters=par, price=price,
                   candidateFile="../inst/extdata/vaccinesSummary.csv")          
}
