% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{portfolioPriceTaker}
\alias{portfolioPriceTaker}
\title{Optimal portfolio for a price taker}
\usage{
portfolioPriceTaker(
  parameters = NULL,
  population,
  gdp_pc,
  frac_high_risk,
  loss2yr,
  price,
  steps = c(10, 1, 0.1),
  candidateFile = NULL,
  lambda = 1,
  return_benefit_args = FALSE
)
}
\arguments{
\item{parameters}{\code{Parameters} object with model parameters}

\item{population}{Country population (in millions)}

\item{gdp_pc}{Country GDP per capita (in thousand $)}

\item{frac_high_risk}{Fraction of population that is high risk}

\item{loss2yr}{Cumulative percent of GDP lost because of pandemic over two years}

\item{price}{Market price for manufacturing capacity ($ per course / year)}

\item{steps}{Steps to optimize over. By default three step-sizes are used in succession:
10, 1 and 0.1.}

\item{candidateFile}{File with candidate data}

\item{return_benefit_args}{if \code{TRUE}, then an extra object is returned, that can then be used
as input to \code{\link[=countryNetBenefits]{countryNetBenefits()}}}
}
\value{
List with information on optimal portfolio. Includes investment
in every candidate, benefits, cost, and total capacity.
}
\description{
Computes the optimal portfolio for a price-taking country.
This is done by calling optimisation function \code{\link[=optimizeGrid]{optimizeGrid()}}
by optimising using decreasing step sizes (see \code{step})
on benefits function \code{\link[=countryNetBenefits]{countryNetBenefits()}}
}
